Update ABCFP to be more self-contained.
The original structure was to have a common directory for
copybooks and a common directory for datafiles.

While there will be some duplication, it seems better to make
each area more independent.

A cpy and a data directory has been added and the files
and datafiles have been moved to those directories.

The code and scripts to compile and execute the programs
have also been updated to point to the new location.